Allow globbing of the 'clean' task.
Instead of prepending `cwd` to each `src` item, the `cwd` should be passed as an option to the `gulp.src()` method.
This way the globbing patterns like `!` are preserved even if `cwd` is provided.
We can use the same method of adding `cwd` to the `src` paths in the `del` function in the `clean` task.